LFOPTEE-293 drivers: caam: Add AES GCM
Implement CAAM AES GCM

CAAM Errata:
When running GCM (not GMAC) when the IV is not 12 bytes (96 bits),
it is possible to "roll over" the 32-bit counter value.
When this occurs (unless counter starts at -1),
the AES operation will generate an incorrect MAC.
This occurs even when -1 is used as the counter for the last block.
The problem is that the 32-bit counter will overflow into the h value,
thus corrupting the MAC.
In order to reliably work around this (unlikely) issue,
the AES operation must be stopped after initialization to
determine the counter value to see whether/when it will roll over.
Then, before the offending block is processed,
the context needs to be saved. The one block gets processed twice :
GCM, restore MAC, GMAC over its ciphertext,
then patch up the message and AAD lengths, and carry on as normal.

Signed-off-by: Olivier Masse <olivier.masse@nxp.com>