core: disable CFG_NOTIF_TEST_WD
CFG_NOTIF_TEST_WD is newly introduced flag in upstream to test
Notification Watchdog.
This depends on CFG_ENABLE_EMBEDDED_TESTS which is enabled by default.
But it also depends on CFG_CORE_ASYNC_NOTIF & CFG_CALLOUT, which is
not enabled on our platforms, which leads to compilation warning.
Since we are not using this feature as of now, disabling it by default.

Signed-off-by: Sahil Malhotra <sahil.malhotra@nxp.com>